Rename hompage folder to homepage
