+++ next/src/pages/index.tsx
@@ -6,10 +6,10 @@ import React from 'react'
 
 import { LATEST_ARTICLES_COUNT } from '@/src/components/common/NavBar/NavMenu/NavMenuLatestArticlesList'
 import PageLayout from '@/src/components/layout/PageLayout'
-import ArticlesHomepageSection from '@/src/components/sections/hompage/ArticlesHomepageSection'
-import HeroHomepageSection from '@/src/components/sections/hompage/HeroHomepageSection'
-import KoloHomepageSection from '@/src/components/sections/hompage/KoloHomepageSection'
-import ServicesHomepageSection from '@/src/components/sections/hompage/ServicesHomepageSection'
+import ArticlesHomepageSection from '@/src/components/sections/homepage/ArticlesHomepageSection'
+import HeroHomepageSection from '@/src/components/sections/homepage/HeroHomepageSection'
+import KoloHomepageSection from '@/src/components/sections/homepage/KoloHomepageSection'
+import ServicesHomepageSection from '@/src/components/sections/homepage/ServicesHomepageSection'
 import { GeneralContextProvider } from '@/src/providers/GeneralContextProvider'
 import { client } from '@/src/services/graphql'
 import { GeneralQuery, HomepageEntityFragment } from '@/src/services/graphql/api'
